Update README.md
5/5 – (4 votes)

In order to nd a single source shortest path, you are asked to implement Dijkstra’s algorithm we went through in class (the PowerPoint Ch12 part2.pptx is attached in this assignment).

Array version

The array version of Dijkstra’s algorithm can be found on slides 23 and 24 in the lecture notes. If you recall, this implementation runs in O(V 2 +E) time, where V is the number of vertices in the graph and E is the number of edges. The O(V 2) component refers to the number of operations carried out to nd every minimum distance vertex extracted from the set of unvisited vertices at the beginning of each iteration of the while loop. The O(E) component results from comparing and possibly updating the distances of all vertices adjacent to the minimum distance vertices. In other words, the for loop within the while loop runs O(E) operations in total.